backend/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
pymongo==4.13.2
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
requests==2.32.4
pandas==2.3.0
numpy==2.2.6
waitress==3.0.2
python-dotenv==1.1.1
flask-jwt-extended==4.7.1
sqlalchemy==2.0.41
scikit-learn==1.7.0
scipy==1.16.0
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.2.0
kaleido==1.0.0
flask-cors==6.0.1